requirements.txt
blinker==1.9.0
cachetools==5.5.0
click==8.1.7
dnspython==1.16.0
Flask==3.1.0
Flask-Cors==5.0.0
google-auth==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.0
pymongo==3.12.0
python-dotenv==1.0.1
rsa==4.9
Werkzeug==3.1.3
